WOAH!
I can see the google stuff! Although there’s still a problem - no such
thing as findOrCreate. Basically, the ‘user’ in the passport docs was
the equivalent of my model.user - and instead of findOrCreate, I need
to use similar code to my login section (create a cookie and spot in
redis for it).